pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dnspython>=2.7.0",
    "email-validator>=2.2.0",
    "pandas>=2.3.2",
    "streamlit>=1.48.1",
]
